+++ defaults.py
@@ -416,7 +416,7 @@ class FlatCAMDefaults:
         "tools_mill_offset_type": 0,  # _('Path')
         "tools_mill_offset_value": 0.0,
         "tools_mill_job_type": 0,  # 'Roughing'
-        "tools_mill_shape": 0,  # 'C1'
+        "tools_mill_tool_shape": 0,  # 'C1'
 
         "tools_mill_cutz": -2.4,
         "tools_mill_vtipdia": 0.1,